backend/requirements.txt
celery
flask
flask-cors
python-dotenv

